[clang-c-frontend] track whether an in-/complete type is required
Fixes #637.

get_type() in the frontend previously always assumed that a complete definition
of the type should be put into the context. As this leads to problems with
mutually recursive types (see #637), a stop-gap measure was in place: put the
(incomplete, but not marked as such!) type in the context and let recursive
calls for its members refer to the wrongly butchered type with no members.
This obviously lead to the types not being equivalent.

Change this by tracking explicitely whether a complete type definition is
required. If not, (as is the case for pointers), still record the incomplete
type in the context, but mark it as such and generate "incomplete_struct" or
"incomplete_union" types instead of the previously butchered "struct" or
"union", respectively.

Every call to get_type() now has to be explicit about whether a complete type
is required to be recorded (for nearly all cases: yes) or not (for pointers).