keep more types symbolic
Together with __ESBMC_ODR-override, this allows the user code to specify
pthread-types that will get replaced by the overridden ones from our c2goto
models.

In particular, the new regression test defines pthread_t and pthread_mutex_t
incompatibly but with size equal or larger to the model's ones. This is still
required! However, all the used pthread_*() functions as well as the mutex and
the thread array now correspond to the model's.

Notably, the constant PTHREAD_MUTEX_INITIALIZER does not work, yet, as the
clang-c-frontend assigns all (named) members (with zero). So the initializer
would be wrong and at this stage there is nothing much we can do as the
ODR-override from the c2goto libs is not even loaded, yet.